wayland-client: fix invalid fd about destroyed proxy
If server sends fds but the proxy was destroyed,
then that fds remain in connetion fds_in array.
so they can be misused by other proxy.

The simple solution is to separate the message including fds from the other messages
and In queueing process in clients, if received message of proxy is invaild,
just consume the fds_in array.

Change-Id: I32d17bcca98c78a60c20618fef9f7ea5eaea7f68
Signed-off-by: Sangjin Lee <lsj119@samsung.com>