connection: send the closure's fds as a group.
Ideally, we can design a request or an event protocol to send several fds like below.

  <request name="send">
    <arg name="first" type="fd"/>
    <arg name="second" type="fd"/>
    ...
  </request>

In this case, we have to send all fds as a group. Therefore, if the number of
fds in a group is larger than the fds_out buffer's remain size, we need to flush
the connection.

Signed-off-by: Boram Park <boram1288.park@samsung.com>