wayland-cursor: fix cursor image broken problem
This fixes a bug related to recreate/use a new shm pool for cursor(s).
When recreating a shm pool, we usually need to copy the existing contents
in the old shm pool to the new shm pool. Otherwise, cursor images in the
old shm pool will be lost and they will be displayed abnormaly.

Change-Id: Id7515f9d15d8406c3dfb1ad52072a2442ce0c1aa
Signed-off-by: Sung-Jin Park <sj76.park@samsung.com>