protocol: add destructor to the registry
In some user side library, it create wl_registry for check global resource to many times.
it also destory the wl_registry when finished but in wayland server, the registry resource remained
untill client is destroy. So memory in server side is increase.
This is the reason why add destructor to wl_registry.

Change-Id: I473e70eb3ef60e6604be4cc0314f3ef95f1f59c4